refactor: split out helpers into related types
