+++ apps/nextjs/src/components/AppComponents/Chat/guidance-required.tsx
@@ -1,7 +1,5 @@
-import {
-  isSafe,
-  moderationGuidanceText,
-} from "@oakai/core/src/utils/ailaModeration/helpers";
+import { moderationGuidanceText } from "@oakai/core/src/utils/ailaModeration/guidanceText";
+import { isSafe } from "@oakai/core/src/utils/ailaModeration/safetyResult";
 import type { PersistedModerationBase } from "@oakai/core/src/utils/ailaModeration/moderationSchema";
 
 import { Icon } from "@/components/Icon";
